Rename json serializer
QRS v1 json serializer is going to be kept for compatibility only and it's
going to be used QJson library instead of QtCore JSON classes since uint64
compatibility problems. Hence JsonSerializer class is going to be renamed
to QJsonSerializer. This commit is first step. It renames files but keep
old class name.